Moving credentials to a separate package
Bitbucket client factory and all the other classes in client package
should not have a source code dependency on jenkins related classes.
Moving credentials to a package helps in breaking this dependency and
moving all logic related to credential handling in one place. One of the
positive side affect of this refactoring is readable test classes.