Remove click to add instance
It seems this breaks on java 11, but works on java 8. It is a strange one. Under java 11 the test will submit two objets in the JSON array,
one with the provided details, and one that has empty values for all keys. It seems as if there was an additional click to add a second
BbS server. Removing the click in the test removes the additional server and the test passes, and importantly it does not break under java 8.
Why this is a problem all of a sudden is a mystery for a different day.