JENKINS-62003 add trigger
Add a new trigger for multibranch. It comes with a few gotchas, first it seems that the trigger is started *before* the scmsource is persisted so any updates to the scm source is not available when the trigger is started.

There seems to be a bug in our Javascript so if you add multiple BbS scm sources it will remove the repository field and thus make it unable to save or function.

Lastly the testing here is very minimal as there are final methods that would need to be mocked in order to construct a proper test. In all a full functional test should be written, but it is not done yet