JENKINS-60116: Allow 'EXTENDED_READ' permission to select credentials
Although JENKINS-60116 already has been closed in November 2019, the faulted
behaviour has not actually been fixed yet as reported by users in that ticket
this year and reproduced by myself.

The method `BitbucketSCMSource.DescriptorImpl.doFillCredentialsIdItems` was
obviously still doing a check for the global 'CONFIGURE' permissions (which
already explains the wrong behaviour and is redundant to the checks in the
called method `BitbucketScmFormFillDelegate.doFillCredentialsIdItems` anyway).

This commit adds the wanted behaviour of a context (item) based permission
check. If a context is given, only the `Item.EXTENDED_READ` permission is
required which follows the approach used by other SCM plugins (Git, Subversion).
Furthermore, the redundant checks are removed and the tests have been adjusted,
to use a mock of the now necessary context parameter.