Remove the PR store.
We do not have the capacity to properly review and test the PullRequestStore right now. Since we plan to release a version of "Build on PR" that does not strictly require it, the store is removed. There is a branch prStoreTrait that preserves this code so we do not loose it and can thus test it properly before release.