Refactor the webhook registration code
This code had become quite hard to reason about when we added the PR support
It came with a number of assumptions that once made sense but in a world
with multiple events and permutations of said events became very difficult
to maintain in a correct form. Thus it was greatly simplified and due to that
the assumption of one event per webhook had to be broken.