Keyhole surgery to support full applinks
An exploration of using what amounts to key-hole surgery to support a full applink. A change is also needed in Bitbucket to support full applinking. This code is an exploration and should not be used, nor released in product, instead consider it an inspiration for how it could possibly be done.